+++ crates/ffi/src/apps/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod close;
+pub(crate) mod launch;
+pub(crate) mod list;
