+++ crates/ffi/src/convert/mod.rs
@@ -4,7 +4,6 @@ pub(crate) mod string;
 pub(crate) mod surface;
 pub(crate) mod window;
 
-pub(crate) use app::{app_info_to_c, free_app_info_fields};
 pub(crate) use rect::rect_to_c;
 pub(crate) use string::{c_to_str, free_c_string, string_to_c};
 pub(crate) use surface::{free_surface_info_fields, surface_info_to_c};
